Italic email doesn't break italicizing other words
This commit fixes a bug that caused strings with multiple lines that are italicized
with an email causes the italic styling to break. With this fix, italicization is applied
like normal except for the email. The email will contain the underscore tags in the
prefix and suffix of the email.