Reduce code regex's to just 1
The code block regex has an escaping and an unescaping version. The unescaping version was transforming the String
to UTF16 and then converting it back to a String which can be very costly because it has to traverse the whole string.
This change is being made to reduce the load time of content.